Struct wallet713::common::crypto::Signature

pub struct Signature(pub(crate) Signature);

An ECDSA signature

Tuple Fields
0: Signature
Implementations

impl Signature

pub fn from_der(secp: &Secp256k1, data: &[u8]) -> Result<Signature, Error>

Converts a DER-encoded byte slice to a signature

pub fn from_compact(secp: &Secp256k1, data: &[u8]) -> Result<Signature, Error>

Converts a 64-byte compact-encoded byte slice to a signature

pub fn from_raw_data(data: &[u8; 64]) -> Result<Signature, Error>

Stores raw bytes provided as a signature, with no conversion

pub fn from_der_lax(secp: &Secp256k1, data: &[u8]) -> Result<Signature, Error>

Converts a “lax DER”-encoded byte slice to a signature. This is basically only useful for validating signatures in the Bitcoin blockchain from before 2016. It should never be used in new applications. This library does not support serializing to this “format”

pub fn normalize_s(&mut self, secp: &Secp256k1)

Normalizes a signature to a “low S” form. In ECDSA, signatures are of the form (r, s) where r and s are numbers lying in some finite field. The verification equation will pass for (r, s) iff it passes for (r, -s), so it is possible to ``modify’’ signatures in transit by flipping the sign of s. This does not constitute a forgery since the signed message still cannot be changed, but for some applications, changing even the signature itself can be a problem. Such applications require a “strong signature”. It is believed that ECDSA is a strong signature except for this ambiguity in the sign of s, so to accomodate these applications libsecp256k1 will only accept signatures for which s is in the lower half of the field range. This eliminates the ambiguity.

However, for some systems, signatures with high s-values are considered valid. (For example, parsing the historic Bitcoin blockchain requires this.) For these applications we provide this normalization function, which ensures that the s value lies in the lower half of its range.

pub fn as_ptr(&self) -> *const Signature

Obtains a raw pointer suitable for use with FFI functions

pub fn as_mut_ptr(&mut self) -> *mut Signature

Obtains a raw mutable pointer suitable for use with FFI functions

pub fn serialize_der(&self, secp: &Secp256k1) -> Vec<u8, Global>

Serializes the signature in DER format

pub fn serialize_compact(&self, secp: &Secp256k1) -> [u8; 64]

Serializes the signature in compact format

pub fn to_raw_data(&self) -> [u8; 64]

Just return raw data, no conversion tricks