Module wallet713::internal::restore

Functions to restore a wallet’s outputs from just the master seed

Structs

OutputResult

Utility struct for return values from below

RestoredTxStats

Collect stats in case we want to just output a single tx log entry for restored non-coinbase outputs

Functions

cancel_tx_log_entry

check_repair

Check / repair wallet contents assume wallet contents have been freshly updated with contents of latest block

collect_chain_outputs

identify_utxo_outputs

restore

Restore a wallet

restore_missing_output