Function wallet713::internal::selection::build_send_tx

pub fn build_send_tx<T: ?Sized, C, K>(
    wallet: &mut T, 
    slate: &mut Slate, 
    minimum_confirmations: u64, 
    max_outputs: usize, 
    change_outputs: usize, 
    selection_strategy_is_use_all: bool, 
    parent_key_id: Identifier
) -> Result<Context, Error> where
    T: WalletBackend<C, K>,
    C: NodeClient,
    K: Keychain, 

Initialize a transaction on the sender side, returns a corresponding libwallet transaction slate with the appropriate inputs selected, and saves the private wallet identifiers of our selected outputs into our transaction context