Module wallet713::internal::updater

Functions

apply_api_outputs

Apply refreshed API output data to the wallet

build_coinbase

Build a coinbase output and insert into wallet

cancel_tx_and_outputs

Cancel transaction and associated outputs

clean_old_unconfirmed

map_wallet_outputs

build a local map of wallet outputs keyed by commit and a list of outputs we want to query the node for

node_height

Fetch the height of the node If the node is down, use the height of the latest output

receive_coinbase

Build a coinbase output and the corresponding kernel

refresh_output_state

Builds a single api query to retrieve the latest output data from the node. So we can refresh the local wallet outputs.

refresh_outputs

Refreshes the outputs in a wallet with the latest information from a node

retrieve_info

Retrieve summary info about the wallet caller should refresh first if desired

retrieve_outputs

Retrieve all of the outputs (doesn’t attempt to update from node)

retrieve_txs

Retrieve all of the transaction entries, or a particular entry if parent_key_id is set, only return entries from that key