Struct wallet713::wallet::api::owner::Owner

pub struct Owner<W, C, K> where
    W: WalletBackend<C, K>,
    C: NodeClient,
    K: Keychain,  {
    container: Arc<Mutex<Container<W, C, K>>>,
}

Fields
container: Arc<Mutex<Container<W, C, K>>>
Implementations

impl<W, C, K> Owner<W, C, K> where
W: WalletBackend<C, K>,
C: NodeClient,
K: Keychain,

pub fn new(container: Arc<Mutex<Container<W, C, K>>>) -> Self

pub fn has_seed(&self) -> Result<bool, Error>

pub fn get_seed(&self) -> Result<ZeroingString, Error>

pub fn set_seed(
&self,
mnemonic: Option<ZeroingString>,
password: ZeroingString,
overwrite: bool
) -> Result<(), Error>

pub fn set_password(&self, password: ZeroingString) -> Result<(), Error>

Set the password to attempt to decrypt the seed with

pub fn connect(&self) -> Result<(), Error>

Connect to the backend

pub fn disconnect(&self) -> Result<(), Error>

Connect to the backend

pub fn clear(&self) -> Result<(), Error>

Clear the wallet of its contents

pub fn config(&self) -> Wallet713Config

pub fn start_listener(
&self,
interface: ListenerInterface
) -> Result<String, Error>

pub fn stop_listener(&self, interface: ListenerInterface) -> Result<bool, Error>

pub fn stop_listeners(&self) -> Result<HashSet<ListenerInterface>, Error>

Stop all running listeners

pub fn epicbox_address(&self) -> Result<EpicboxAddress, Error>

pub fn set_epicbox_address_index(
&self,
index: u32
) -> Result<EpicboxAddress, Error>

pub fn accounts(&self) -> Result<Vec<AcctPathMapping>, Error>

pub fn create_account_path(&self, label: &str) -> Result<Identifier, Error>

pub fn active_account(&self) -> Result<String, Error>

pub fn set_active_account(&self, label: &str) -> Result<(), Error>

pub fn contacts(&self) -> Result<Vec<Contact>, Error>

pub fn add_contact(&self, name: &str, address: &str) -> Result<(), Error>

pub fn remove_contact(&self, name: &str) -> Result<(), Error>

pub fn retrieve_outputs(
&self,
include_spent: bool,
refresh_from_node: bool,
tx_id: Option<u32>
) -> Result<(bool, Option<u64>, Vec<OutputCommitMapping>), Error>

pub fn retrieve_txs(
&self,
refresh_from_node: bool,
check_contacts: bool,
check_proofs: bool,
tx_id: Option<u32>,
tx_slate_id: Option<Uuid>
) -> Result<(bool, Option<u64>, Vec<TxLogEntry>, HashMap<String, String>, HashMap<Uuid, bool>), Error>

fn retrieve_tx(
&self,
tx_id: Option<u32>,
tx_slate_id: Option<Uuid>
) -> Result<TxLogEntry, Error>

pub fn retrieve_summary_info(
&self,
refresh_from_node: bool,
minimum_confirmations: u64
) -> Result<(bool, WalletInfo), Error>

pub fn init_send_tx(&self, args: InitTxArgs) -> Result<Slate, Error>

pub fn tx_lock_outputs(
&self,
slate: &Slate,
participant_id: usize,
address: Option<String>
) -> Result<(), Error>

pub fn finalize_tx(
&self,
slate: &Slate,
tx_proof: Option<&mut TxProof>
) -> Result<Slate, Error>

pub fn post_tx(&self, tx: &Transaction, fluff: bool) -> Result<(), Error>

pub fn cancel_tx(
&self,
tx_id: Option<u32>,
tx_slate_id: Option<Uuid>
) -> Result<(), Error>

pub fn get_stored_tx(
&self,
slate_id: &Uuid
) -> Result<Option<Transaction>, Error>

pub fn repost_tx(
&self,
tx_id: Option<u32>,
tx_slate_id: Option<Uuid>,
fluff: bool
) -> Result<Uuid, Error>

pub fn verify_slate_messages(&self, slate: &Slate) -> Result<(), Error>

pub fn get_stored_tx_proof(
&self,
tx_id: Option<u32>,
tx_slate_id: Option<Uuid>
) -> Result<Option<TxProof>, Error>

pub fn verify_tx_proof(
&self,
tx_proof: &TxProof
) -> Result<(EpicboxAddress, EpicboxAddress, u64, Vec<Commitment>, Commitment), Error>

pub fn restore(&self) -> Result<(), Error>

pub fn check_repair(&self, delete_unconfirmed: bool) -> Result<(), Error>

pub fn node_height(&self) -> Result<NodeHeightResult, Error>

pub fn node_version(&self) -> Option<NodeVersionInfo>

fn open_and_close<F, X>(&self, f: F) -> Result<X, Error> where
F: FnOnce(&mut MutexGuard<'_, Container<W, C, K>>) -> Result<X, Error>,

Convenience function that opens and closes the wallet with the stored credentials