Module wallet713::wallet::types

Re-exports

pub use self::acct_path_mapping::AcctPathMapping;

pub use self::args::*;

pub use self::block_fees::BlockFees;

pub use self::block_identifier::BlockIdentifier;

pub use self::cb_data::CbData;

pub use self::context::Context;

pub use self::node_client::HTTPNodeClient;

pub use self::node_client::NodeClient;

pub use self::node_client::NodeVersionInfo;

pub use self::output_commit_mapping::OutputCommitMapping;

pub use self::output_data::OutputData;

pub use self::output_status::OutputStatus;

pub use self::slate::Slate;

pub use self::slate::SlateVersion;

pub use self::slate::VersionedSlate;

pub use self::slate::CURRENT_SLATE_VERSION;

pub use self::slate::EPIC_BLOCK_HEADER_VERSION;

pub use self::tx_log_entry::TxLogEntry;

pub use self::tx_log_entry_type::TxLogEntryType;

pub use self::tx_proof::ErrorKind as TxProofErrorKind;

pub use self::tx_proof::TxProof;

pub use self::tx_wrapper::TxWrapper;

pub use self::wallet_backend::WalletBackend;

pub use self::wallet_backend_batch::WalletBackendBatch;

pub use self::wallet_info::WalletInfo;

pub use self::wallet_inst::WalletInst;

pub use super::seed::EncryptedWalletSeed;

pub use super::seed::WalletSeed;

pub use crate::common::Result;

Modules

acct_path_mapping

args

block_fees

block_identifier

cb_data

context

node_client

output_commit_mapping

output_data

output_status

slate

tx_log_entry

tx_log_entry_type

tx_proof

tx_wrapper

wallet_backend

wallet_backend_batch

wallet_info

wallet_inst

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

ExtKeychain

Hash

A hash to uniquely (or close enough) identify one of the main blockchain constructs. Used pervasively for blocks, transactions and outputs.

Identifier

Output

Output for a transaction, defining the new ownership of coins that are being transferred. The commitment is a blinded value for the output while the range proof guarantees the commitment includes a positive value without overflow and the ownership of the private key.

PublicKey

A Secp256k1 public key, used for verification of signatures

SecretKey

Secret 256-bit key used as x in an ECDSA signature

Transaction

A transaction

TxKernel

A proof that a transaction sums to zero. Includes both the transaction’s Pedersen commitment and the signature, that guarantees that the commitments amount to zero. The signature signs the fee and the lock_height, which are retained for signature validation.

Enums

ChildNumber

A child number for a derived key

Traits

Keychain

Type Definitions

Mutex

A mutual exclusion primitive useful for protecting shared data

MutexGuard

An RAII implementation of a “scoped lock” of a mutex. When this structure is dropped (falls out of scope), the lock will be unlocked.