Module wallet713::wallet::types::slate::slate

Functions for building partial transactions to be passed around during an interactive wallet exchange

Structs

ParticipantData

Public data for each participant in the slate

ParticipantMessageData

Public message data (for serialising and storage)

ParticipantMessages

Helper just to facilitate serialization

Slate

A ‘Slate’ is passed around to all parties to build up all of the public transaction data needed to create a finalized transaction. Callers can pass the slate around by whatever means they choose, (but we can provide some binary or JSON serialization helpers here).

VersionCompatInfo

Versioning and compatibility info about this slate