Module wallet713::wallet::types::slate::versions::v2

Contains V2 of the slate (epic-wallet 1.1.0) Changes from V1:

ParticipantData struct fields serialized as hex strings instead of arrays:

public_blind_excess

public_nonce

part_sig

message_sig

Transaction fields serialized as hex strings instead of arrays:

offset

Input field serialized as hex strings instead of arrays: commit

Output fields serialized as hex strings instead of arrays: commit proof

TxKernel fields serialized as hex strings instead of arrays: commit signature

version field removed

VersionCompatInfo struct created with fields and added to beginning of struct version: u16 orig_verion: u16, block_header_version: u16

Structs

InputV2

OutputV2

ParticipantDataV2

SlateV2

TransactionBodyV2

TransactionBody is a common abstraction for transaction and block

TransactionV2

A transaction

TxKernelV2

VersionCompatInfoV2