Struct wallet713::wallet::types::Transaction

pub struct Transaction {
    pub offset: BlindingFactor,
    pub body: TransactionBody,
}

A transaction

Fields
offset: BlindingFactor

The kernel “offset” k2 excess is k1G after splitting the key k = k1 + k2

body: TransactionBody

The transaction body - inputs/outputs/kernels

Implementations

impl Transaction

pub fn empty() -> Transaction

Creates a new empty transaction (no inputs or outputs, zero fee).

pub fn new(
inputs: Vec<Input, Global>,
outputs: Vec<Output, Global>,
kernels: Vec<TxKernel, Global>
) -> Transaction

Creates a new transaction initialized with the provided inputs, outputs, kernels

pub fn with_offset(self, offset: BlindingFactor) -> Transaction

Creates a new transaction using this transaction as a template and with the specified offset.

pub fn with_input(self, input: Input) -> Transaction

Builds a new transaction with the provided inputs added. Existing inputs, if any, are kept intact. Sort order is maintained.

pub fn with_output(self, output: Output) -> Transaction

Builds a new transaction with the provided output added. Existing outputs, if any, are kept intact. Sort order is maintained.

pub fn with_kernel(self, kernel: TxKernel) -> Transaction

Builds a new transaction with the provided kernel added. Existing kernels, if any, are kept intact. Sort order is maintained.

pub fn replace_kernel(self, kernel: TxKernel) -> Transaction

Builds a new transaction replacing any existing kernels with the provided kernel.

pub fn inputs(&self) -> &Vec<Input, Global>

Get inputs

pub fn inputs_mut(&mut self) -> &mut Vec<Input, Global>

Get inputs mutable

pub fn outputs(&self) -> &Vec<Output, Global>

Get outputs

pub fn outputs_mut(&mut self) -> &mut Vec<Output, Global>

Get outputs mutable

pub fn kernels(&self) -> &Vec<TxKernel, Global>

Get kernels

pub fn kernels_mut(&mut self) -> &mut Vec<TxKernel, Global>

Get kernels mut

pub fn fee(&self) -> u64

Total fee for a transaction is the sum of fees of all kernels.

pub fn overage(&self) -> i64

Total overage across all kernels.

pub fn lock_height(&self) -> u64

Lock height of a transaction is the max lock height of the kernels.

pub fn validate_read(&self) -> Result<(), Error>

“Lightweight” validation that we can perform quickly during read/deserialization. Subset of full validation that skips expensive verification steps, specifically -

rangeproof verification (on the body)

kernel signature verification (on the body)

kernel sum verification

pub fn validate(&self, weighting: Weighting) -> Result<(), Error>

Validates all relevant parts of a fully built transaction. Checks the excess value against the signature as well as range proofs for each output.

pub fn fee_to_weight(&self) -> u64

Can be used to compare txs by their fee/weight ratio. Don’t use these values for anything else though due to precision multiplier.

pub fn tx_weight(&self) -> usize

Calculate transaction weight

pub fn tx_weight_as_block(&self) -> usize

Calculate transaction weight as a block

pub fn weight(input_len: usize, output_len: usize, kernel_len: usize) -> usize

Calculate transaction weight from transaction details