Trait wallet713::wallet::types::wallet_backend::WalletBackend

pub trait WalletBackend<C, K>: Send + 'static where
    C: NodeClient,
    K: Keychain, {

    fn has_seed(&self) -> StdResult<bool, Error>;

    fn get_seed(&self) -> StdResult<ZeroingString, Error>;

    fn set_seed(
        &mut self, 
        mnemonic: Option<ZeroingString>, 
        password: ZeroingString, 
        overwrite: bool
    ) -> StdResult<(), Error>;

    fn connected(&self) -> StdResult<bool, Error>;

    fn connect(&mut self) -> StdResult<(), Error>;

    fn disconnect(&mut self) -> StdResult<(), Error>;

    fn set_password(&mut self, password: ZeroingString) -> StdResult<(), Error>;

    fn clear(&mut self) -> StdResult<(), Error>;

    fn open_with_credentials(&mut self) -> StdResult<(), Error>;

    fn close(&mut self) -> StdResult<(), Error>;

    fn restore(&mut self) -> StdResult<(), Error>;

    fn check_repair(&mut self, delete_unconfirmed: bool) -> StdResult<(), Error>;

    fn get_parent_key_id(&self) -> Identifier;

    fn set_parent_key_id(&mut self, id: &Identifier);

    fn set_parent_key_id_by_name(&mut self, label: &str) -> StdResult<(), Error>;

    fn w2n_client(&mut self) -> &mut C;

    fn calc_commit_for_cache(
        &mut self, 
        amount: u64, 
        id: &Identifier
    ) -> StdResult<Option<String>, Error>;

    fn keychain(&mut self) -> &mut K;

    fn next_child(&mut self) -> StdResult<Identifier, Error>;

    fn get_output(
        &self, 
        id: &Identifier, 
        mmr_index: &Option<u64>
    ) -> StdResult<OutputData, Error>;

    fn get_private_context(
        &mut self, 
        slate_id: &[u8], 
        participant_id: usize
    ) -> StdResult<Context, Error>;

    fn get_acct_path(
        &self, 
        label: &str
    ) -> StdResult<Option<AcctPathMapping>, Error>;

    fn get_last_confirmed_height(&self) -> StdResult<u64, Error>;

    fn get_stored_tx(&self, uuid: &str) -> StdResult<Option<Transaction>, Error>;

    fn has_stored_tx_proof(&self, uuid: &str) -> StdResult<bool, Error>;

    fn get_stored_tx_proof(
        &self, 
        uuid: &str
    ) -> StdResult<Option<TxProof>, Error>;

    fn get_tx_log_by_slate_id(
        &self, 
        slate_id: &str
    ) -> StdResult<Option<TxLogEntry>, Error>;

    fn outputs<'a>(
        &'a self
    ) -> StdResult<Box<dyn Iterator<Item = OutputData> + 'a>, Error>;

    fn tx_logs<'a>(
        &'a self
    ) -> StdResult<Box<dyn Iterator<Item = TxLogEntry> + 'a>, Error>;

    fn accounts<'a>(
        &'a self
    ) -> StdResult<Box<dyn Iterator<Item = AcctPathMapping> + 'a>, Error>;

    fn batch<'a>(
        &'a self
    ) -> StdResult<Box<dyn WalletBackendBatch<K> + 'a>, Error>;

}

Required methods

fn has_seed(&self) -> StdResult<bool, Error>

Check whether the backend has a seed or not

fn get_seed(&self) -> StdResult<ZeroingString, Error>

Get the seed

fn set_seed(
&mut self,
mnemonic: Option<ZeroingString>,
password: ZeroingString,
overwrite: bool
) -> StdResult<(), Error>

Set a new seed, encrypt with password Should fail if backend already has a seed, unless overwrite is set to `true

fn connected(&self) -> StdResult<bool, Error>

Check if the backend connection is established

fn connect(&mut self) -> StdResult<(), Error>

Connect to the backend

fn disconnect(&mut self) -> StdResult<(), Error>

Disconnect from backend

fn set_password(&mut self, password: ZeroingString) -> StdResult<(), Error>

Set password

fn clear(&mut self) -> StdResult<(), Error>

Clear out backend

fn open_with_credentials(&mut self) -> StdResult<(), Error>

fn close(&mut self) -> StdResult<(), Error>

fn restore(&mut self) -> StdResult<(), Error>

fn check_repair(&mut self, delete_unconfirmed: bool) -> StdResult<(), Error>

fn get_parent_key_id(&self) -> Identifier

fn set_parent_key_id(&mut self, id: &Identifier)

fn set_parent_key_id_by_name(&mut self, label: &str) -> StdResult<(), Error>

fn w2n_client(&mut self) -> &mut C

fn calc_commit_for_cache(
&mut self,
amount: u64,
id: &Identifier
) -> StdResult<Option<String>, Error>

fn keychain(&mut self) -> &mut K

fn next_child(&mut self) -> StdResult<Identifier, Error>

fn get_output(
&self,
id: &Identifier,
mmr_index: &Option<u64>
) -> StdResult<OutputData, Error>

fn get_private_context(
&mut self,
slate_id: &[u8],
participant_id: usize
) -> StdResult<Context, Error>

fn get_acct_path(
&self,
label: &str
) -> StdResult<Option<AcctPathMapping>, Error>

fn get_last_confirmed_height(&self) -> StdResult<u64, Error>

fn get_stored_tx(&self, uuid: &str) -> StdResult<Option<Transaction>, Error>

fn has_stored_tx_proof(&self, uuid: &str) -> StdResult<bool, Error>

fn get_stored_tx_proof(&self, uuid: &str) -> StdResult<Option<TxProof>, Error>

fn get_tx_log_by_slate_id(
&self,
slate_id: &str
) -> StdResult<Option<TxLogEntry>, Error>

fn outputs<'a>(
&'a self
) -> StdResult<Box<dyn Iterator<Item = OutputData> + 'a>, Error>

fn tx_logs<'a>(
&'a self
) -> StdResult<Box<dyn Iterator<Item = TxLogEntry> + 'a>, Error>

fn accounts<'a>(
&'a self
) -> StdResult<Box<dyn Iterator<Item = AcctPathMapping> + 'a>, Error>

fn batch<'a>(&'a self) -> StdResult<Box<dyn WalletBackendBatch<K> + 'a>, Error>